Completable tasks and removable folders
Tasks are now completable. Folders are removable. When the user tries to
remove a folder, he is warned by a confirmation dialog.

The bug where the tasks were cleared each time the app was created is
now fixed.

A message now always tells the user when there are no folders or tasks.
This was displayed wrongly earlier.

reset bug