Removable keyboard
The keyboard, which was previously unremovable in some occasions, should
now be removable when pressing the enter key.

Since the intent.getExtras().getString(name) is not a valid function on
Android 2.2, this was changed to intent.getStringExtra(name).

The save button for the folder and task names should now appear and
disappear correctly.