Improved header
Since something in GitHub went wrong, the entire code that was written
during the last days ad to be rewritten. Some bugs may still exist.

The names of the folders and tasks are displayed through a text view.
When this text view is pressed, it disappears. Instead, an edit text and
a button become visible. The name can then be edited.

In the task activity, an edit text with a width and height of 0 has been
placed. This edit text works as a focus dummy; it takes the unwanted
focus from the other edit texts when it must do so.