Started a new project
The Code2 project is a completely new version of the app. The data is
stored in JSON objects, just like in Code. Some minor graphics have been
implemented. Also, a menu, which is accessible by pressing or dragging
the dark green area at the upper left corener, has been added.

At the moment it is just possible to add tasks in the default Inbox
folder. These can be set to "done" when the X-button is clicked.