Started another new project
The fly in menu used in this code is taken from a library made by Alex
Curran (https://github.com/Espiandev/FlyInMenu).

Completable tasks in the inbox folder have been added. It is also
possible to make new folders and add checklists to them.

By pressing the dark green button at the top left corner of the screen,
the menu is shown.