Reworked the way content is displayed
When I looked at the files that were supposed to be synced to GitHub, I
noticed they were not. For some unknown reason the files were spread out
over my computer and not in the folder being synced to GitHub.
Therefore, the files have not been updated. Everything should be working
properly now though.

Instead of being based on lists, content is now displayed in views
handled by separate classes. This both cleans up the MainActivity class
and makes it easier to in the future view content that is not list based
(such as projects and task descriptions).