Enterable tasks and animation
When a task is clicked, the user can see the tasks content. At the
moment all that is being shown is a text view which is replaced by an
edit text when clicked. Later, when testing on a real device, this will
be made prettier. Descriptions, due dates, categorization and much more
should be editable in the task content views.

The animating effect between the different views should now be working
properly. Some bugs may still remain.

The interpolatur used by the fly in menu is now also used in the
animation between the slides. Using different ways of interpolating the
animation looked terrible.