Added task descriptions
Task descriptions can now be added. These are currently not displayed in
a good way and need to be styled.

The title text field at the top of the screen has now a max width. If a
title is too long, letters will be cut and "..." will be added until the
title fits the text field.