Rebuild the id system
Tasks, checklists and folders now have a unique id . This will make it
easier later, when syncing a users data to a database.

A new property has been added to all the tasks, checklists and folders
(CHILDREN_IDS) which holds information about the children ids. The
property that hold the value of the number of children (NUM_CHILDREN)
has been removed.