Tasks and checklists can change order
If the API level of the device is over 11 (Android 3.0), the user can
drag around tasks and checklists thus changing their order. By holding
the item for a while, it becomes draggable. When the grabbed item is
dragged, the other list items will dynamically be changed.

As the user releases the item, it's new position is saved in the data
string. This means that the order will be the same the next time the
user opens the folder or checklist again.