Enhanced the fly in menu
When the user makes a swiping gesture at the left end of the screen, the
fly in menu becomes draggable; it follows the users finger. When the
finger is released, it is determined whether the menu should open or
close. The velocity of the swipe and the swipe distance is taken into
consideration when determining this.