Clean up and new colors
Some static methods from the MainActivity class have been moved to the
App class. This was done to clean up the MainActivity class and make it
shorter.

The status bar is visible. Also, some of the colors have been changed to
match the chosen color palette.

The bug where the content of the menu became invisible before it was
outside the screen, has been resolved.