Added animations
Some animations have been added. When a task is checked, it is easier
for the user to follow it. It is places at the top of all the other
checked tasks.

Furthermore, when adding tasks and checklists, an animation shows this.

The bug where the menu became black when it was closed with a fling, has
been resolved.