Rebuilt the structure of the app
Data is now saved in a different way. All the items are named after
their id. They now have a type property which gives information about
what type the item is.

Checklists and projects have been removed. Everything has been made
simpler so that using the app will be easier. Furthermore, folders and
tasks (and also notes in the future) can be stored and shown in the same
folder.

A lot of files are at the moment not being used. Later a huge clean-up
will be made to remove all the unnecessary files.