Framework for options for items
When an item is long pressed, options for all items are shown. In
connection with this, an option bar is shown at the bottom of the
screen. It will later be styled and buttons will be placed on it.

All the folders have a checkbox (where the user can select the items).
When the used presses the options bar at the bottom (which at the moment
is empty) the selected items are removed.