New reminders, dynamic folder icons and sorting
Reminders can now be set for a task, either relative to due or absolute.
Repeating reminders have not yet been added.

The folder icons are different if a folder contains a task which is over
due or has all tasks in it completed.

How to sort items can be decided in the settings.

Lots of other small things have also been fixed and added.